Add files for UI tests

+++ src/test/java/com/example/teamcity/BaseTest.java
@@ -4,9 +4,14 @@ import com.example.teamcity.api.generators.TestDataStorage;
 import com.example.teamcity.api.models.TestData;
 import com.example.teamcity.api.requests.CheckedRequests;
 import com.example.teamcity.api.spec.Specifications;
+import org.testng.ITestResult;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.Test;
 import org.testng.asserts.SoftAssert;
+
+import java.lang.reflect.Method;
+import java.util.Arrays;
 
 import static com.example.teamcity.api.generators.TestDataGenerator.generate;
 
@@ -23,8 +28,15 @@ public class BaseTest {
     }
 
     @AfterMethod(alwaysRun = true)
-    public void afterTest() {
+    public void afterTest(ITestResult result) {
         softy.assertAll();
-        TestDataStorage.getStorage().deleteCreatedEntities();
+        boolean isUiTest = Arrays.asList(result.getMethod().getGroups()).contains("UI");
+
+        if (!isUiTest) {
+            System.out.println("Cleaning up after API test...");
+            TestDataStorage.getStorage().deleteCreatedEntities();
+        } else {
+            System.out.println("Skipping cleanup for UI test.");
+        }
     }
 }
